Changed rectTransform class to stop using two booleans
The offset representation on its own was able to represent position-size scaleings just as well, so I didn't have to seperate the functionalities. A lot of unneeded code was removed as a result. It was a but unintuitive and I didnt see this till I built the editor with the redundant feature, so i guess it wasnt a complete waste of time