Non-typesafe callbacks may now be added to Property<T>
This allows for removing the large if-else in SetupProperty in ComponentEditorUI.cs in favour of a single line of code. It does however result in worse performance. The Action<object> is converted to an Action<T> by creating a new Action<T> that simply calls the Action<object> lambda. In order for adding and removing to work, a dictionary mapping the Action<object> to Action<T> must also be maintained. I have initialized it to null to avoid too much extra memory being allocated for this purpose. Also removed the if-else used to make a variable editor into its own static class